requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
frozenlist==1.4.1
idna==3.10
imageio==2.35.1
multidict==6.1.0
numpy==2.0.2
pandas==2.2.3
pillow==10.4.0
propcache==0.2.0
py-cord==2.6.1
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.14.0
